+++ tests/spec/Gnugat/Redaktilo/Search/LineNumberSearchStrategySpec.php
@@ -12,7 +12,6 @@
 namespace spec\Gnugat\Redaktilo\Search;
 
 use Gnugat\Redaktilo\Converter\LineContentConverter;
-use Gnugat\Redaktilo\FactoryMethod\LineNumber;
 use Gnugat\Redaktilo\File;
 use PhpSpec\ObjectBehavior;
 
@@ -22,7 +21,7 @@ class LineNumberSearchStrategySpec extends ObjectBehavior
 
     function let(File $file, LineContentConverter $converter)
     {
-        $rootPath = __DIR__.'/../../../../../';
+        $rootPath = __DIR__.'/../../../../..';
 
         $filename = sprintf(self::FILENAME, $rootPath);
         $lines = file($filename, FILE_IGNORE_NEW_LINES);
@@ -39,7 +38,7 @@ class LineNumberSearchStrategySpec extends ObjectBehavior
 
     function it_supports_line_numbers()
     {
-        $lineNumber = LineNumber::absolute(42);
+        $lineNumber = 42;
         $line = 'Sir Bedevere: Good. Now, why do witches burn?';
         $rawLine = $line."\n";
 
@@ -48,42 +47,35 @@ class LineNumberSearchStrategySpec extends ObjectBehavior
         $this->supports($rawLine)->shouldBe(false);
     }
 
-    function it_checks_line_presence(File $file)
+    function it_finds_previous_occurences(File $file)
     {
-        $existingLine = LineNumber::absolute(5);
-        $nonExistingLine = LineNumber::absolute(1337);
+        $currentLineNumber = 5;
+        $existingLine = 4;
+        $nextLineNumber = 1;
+        $nonExistingLine = 23;
 
-        $this->has($file, $existingLine)->shouldBe(true);
-        $this->has($file, $nonExistingLine)->shouldBe(false);
+        $this->findPrevious($file, $nonExistingLine, $currentLineNumber)->shouldBe(false);
+        $this->findPrevious($file, $existingLine, $currentLineNumber)->shouldBe($nextLineNumber);
+
+        $file->getCurrentLineNumber()->willReturn($currentLineNumber);
+
+        $this->findPrevious($file, $nonExistingLine)->shouldBe(false);
+        $this->findPrevious($file, $existingLine)->shouldBe($nextLineNumber);
     }
 
     function it_finds_next_occurences(File $file)
     {
-        $currentLineNumber = LineNumber::absolute(5);
-        $existingLine = LineNumber::down(2);
-        $nextLineNumber = LineNumber::absolute(7);
-        $nonExistingLine = LineNumber::down(23);
+        $currentLineNumber = 5;
+        $existingLine = 2;
+        $nextLineNumber = 7;
+        $nonExistingLine = 23;
+
+        $this->findNext($file, $nonExistingLine, $currentLineNumber)->shouldBe(false);
+        $this->findNext($file, $existingLine, $currentLineNumber)->shouldBe($nextLineNumber);
 
         $file->getCurrentLineNumber()->willReturn($currentLineNumber);
 
-        $exception = 'Gnugat\Redaktilo\Search\PatternNotFoundException';
-
-        $this->shouldThrow($exception)->duringFindNext($file, $nonExistingLine);
+        $this->findNext($file, $nonExistingLine)->shouldBe(false);
         $this->findNext($file, $existingLine)->shouldBe($nextLineNumber);
-    }
-
-    function it_finds_previous_occurences(File $file)
-    {
-        $currentLineNumber = LineNumber::absolute(5);
-        $existingLine = LineNumber::up(4);
-        $nextLineNumber = LineNumber::absolute(1);
-        $nonExistingLine = LineNumber::up(23);
-
-        $file->getCurrentLineNumber()->willReturn($currentLineNumber);
-
-        $exception = 'Gnugat\Redaktilo\Search\PatternNotFoundException';
-
-        $this->shouldThrow($exception)->duringFindPrevious($file, $nonExistingLine);
-        $this->findPrevious($file, $existingLine)->shouldBe($nextLineNumber);
     }
 }